refactor: move resolve_fast_field_by_name to shared pullup module
The helper belongs next to resolve_fast_field in pullup.rs rather than
in datafusion_build.rs, since it's a general-purpose fast field
resolution utility.
